Improve build system (#12)
* Improve build system

We use a symbolic link to switch between the two versions of
`package.yaml` for static and dynamic builds, respectively.

This allows us to remove a brittle hack that pathces `package.yaml` at
build time.

* Use YAML include mechanism to include build options

More build system improvements. This way, we don't need to maintain two
versions of the `package.yaml` file.