Downgrade to stack resolver 17.04 (#14)
* Downgrade to stack resolver 17.04

Downgrade resolver to lts 17.4 so we could build using Nix.

* Reproducible static build with Nix

This allows us to determinisitically build `sfsigner` on Linux. The
process of creating a static stack build in Nix is extremely difficult
and lengthy, possibly fragile, but it's not impossible.

The built artifact is in `r10e-static-build/out/sfsigner-r10e`

* Refactor: fix NIX_PATH

A small but important refactor to fix the NIX_PATH.

* More improvements

Add more improvements. We remove the malfunctional filter from the
dockerclean make target.